Simplify model.components
It was a 1x1 cell array wrapping a single struct array, which made it easier to
store things with new fields, since for some reason Matlab doesn't apply the
same rules to structs in cell arrays that it does to ordinary structs. Now it's
just a normal struct (which hopefully won't break things?).